Organizes cart logic into cart component
Places cart code into cart component to clean up header component
